Update linuxcnc_post.py to add --dwell "nn.n" to arguments.
By adding --dwell "n.nn" , where "n.nn" is a decimal number representing seconds, this new processor will export G-code for LinuxCNC machines like the old one did, but allow users to add G4 nn.n to the generated *.nc file. This will be inserted after any M3 Sxxxx spindle start command. The purpose and effect is to give time for the spindle to spin up before things begin: in many situations this is desirable. If the --dwell argument is omitted, then the program will function as it did before.